% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpCycle.R
\name{rpCycle}
\alias{rpCycle}
\title{The cell cycle phase prportion in samples}
\usage{
rpCycle(expr, marker = genemarker)
}
\arguments{
\item{expr}{the gene expression data set. A data frame with row names as symbols and columns as samples.}

\item{marker}{a list of cell cycle phase marker symbol genes.}
}
\value{
a matrix, row is samples, col is cell cycle phase.
}
\description{
calculated the relative proportion of cells in each cell cycle phase for samples based on gene expression profile.
}
